Simplify test for AMP plugins (#6311)
* AMP

* fuse

* yapf
